naturalize fixture additions — drop fp-guards/ subdir, distribute by domain
The first round of this PR added a code/src/fp-guards/ subdirectory with
files named after the test (enum-flow-status-regression.ts,
enum-entity-collision-task-visibility.ts, etc.), including one file that
was entirely comments. The IL test harness scans every .ts under code/src/
recursively and only cares that the verifier's drift set matches the
union of // IL-DRIFT: markers — the *location* of those markers has no
semantic meaning. So the fp-guards/ scaffolding was test-organizing
imposed on top of a fixture project that's supposed to look like a real
e-commerce/catalog application.

Drop the scaffolding. Distribute the same content by application domain:

  fp-guards/enum-entity-collision-visibility.ts
    → WidgetVisibility type joins handlers/widget-handler.ts (widgets
      already live there) as an FP-guard inline with the existing handler
  fp-guards/enum-entity-collision-task-visibility.ts
    → new handlers/task-handler.ts — a believable task-assignment module
      that owns taskVisibility (regression target) + TaskInputSchema
      (the collision target inside a Zod schema)
  fp-guards/enum-flow-status-alias.ts + .../regression.ts
    → new services/sync-pipeline.service.ts — a sync orchestration service
      with RunState (FP-guard) and a top-of-file IL-DRIFT marker for the
      PipelineStage regression, narrated as "stages live elsewhere"
  fp-guards/enum-operation-type-regression.ts
    → IL-DRIFT marker moved into plugins/catalog/server/routes/items.ts
      with a one-line natural comment about lifecycle events
  fp-guards/processors/{email,sms,push,webhook}/index.ts
    → moved to processors/{email,sms,push,webhook}/index.ts — these were
      already domain-named; the fp-guards/ prefix was the only artifice

Inferred contract _inferred/visibility/visibility.tc updated to point at
the new location of the visibility Zod field (task-handler.ts:29).

All 387 contract-verifier tests pass, including the 8 end-to-end fixture
tests that compare the verifier's drift set against the IL-DRIFT markers.
The drift set is unchanged from before this refactor — it's strictly an
organizational change to the fixture project's source layout.
