--- tests/fixtures/sample-js-project-il/code/src/fp-guards/enum-flow-status-regression.ts
@@ -1,4 +1,0 @@
-// Regression: PipelineStage genuinely has no code counterpart — its values
-// do not appear as any named enum or type alias anywhere in the codebase.
-// no-code-counterpart must still fire after the value-set match fix.
-// IL-DRIFT: Enum:PipelineStage / enum.PipelineStage.no-code-counterpart
