--- tests/fixtures/sample-js-project-il/code/src/fp-guards/enum-operation-type-regression.ts
@@ -1,4 +1,0 @@
-// Regression: PluginEventType has no code counterpart — no sibling-module id-literal
-// pattern exists for these values anywhere in the codebase.
-// no-code-counterpart must still fire after the sibling synthesis fix.
-// IL-DRIFT: Enum:PluginEventType / enum.PluginEventType.no-code-counterpart
