+++ tests/fixtures/sample-js-project-il/code/src/plugins/catalog/server/routes/items.ts
@@ -6,6 +6,11 @@
 // The comparator must treat an empty code-side response list as "unverifiable",
 // not "response 200 was never emitted".
 
+// Plugin lifecycle events (`installed`, `activated`, `deactivated`, `removed`)
+// are dispatched by the plugin host rather than declared here; the spec-side
+// `PluginEventType` enum has no code counterpart inside this plugin module.
+// IL-DRIFT: Enum:PluginEventType / enum.PluginEventType.no-code-counterpart
+
 export default [
   {
     method: 'GET',
