+++ tests/fixtures/sample-js-project-il/reference/contracts/_inferred/visibility/visibility.tc
@@ -1,6 +1,6 @@
 enum visibility {
   // inferred — enum defined in code but documented in no spec
-  inferred-from "src/fp-guards/enum-entity-collision-visibility.ts" 14..14
+  inferred-from "src/handlers/task-handler.ts" 29..29
   confidence high
   representation string-literal
   closed
